升级经验更改数据类型为quint64,防止溢出；增加职业加成设定；

+++ code/GenerateSettingFile/lvExp.h
@@ -2,7 +2,7 @@
 
 const qint32 maxLv = 1000;
 
-qint32 expNeed[maxLv] = {
+quint64 expNeed[maxLv] = {
 	100,
 	200,
 	300,
